+++ src/components/scenes/Fio/FioSentRequestDetailsScene.tsx
@@ -6,7 +6,7 @@ import { FIAT_CODES_SYMBOLS } from '../../../constants/WalletAndCurrencyConstant
 import { formatDate, formatNumber, SHORT_DATE_FMT } from '../../../locales/intl'
 import { lstrings } from '../../../locales/strings'
 import { connect } from '../../../types/reactRedux'
-import { EdgeSceneProps } from '../../../types/routerTypes'
+import { EdgeAppSceneProps } from '../../../types/routerTypes'
 import { FioRequest, FioRequestStatus, GuiExchangeRates } from '../../../types/types'
 import { EdgeCard } from '../../cards/EdgeCard'
 import { SceneWrapper } from '../../common/SceneWrapper'
@@ -19,7 +19,7 @@ export interface FioSentRequestDetailsParams {
   selectedFioSentRequest: FioRequest
 }
 
-interface OwnProps extends EdgeSceneProps<'fioSentRequestDetails'> {}
+interface OwnProps extends EdgeAppSceneProps<'fioSentRequestDetails'> {}
 
 interface StateProps {
   fiatSymbol: string